django_testing/requirements.txt
Django==3.1.2
djangorestframework==3.12.1
psycopg2-binary==2.8.6

pytest~=6.1.2
django-filter~=2.4.0